Renamed the renderable
