+++ ugs-platform/ugs-platform-visualizer/src/main/java/com/willwinder/ugs/nbm/visualizer/shared/GcodeRenderer.java
@@ -25,7 +25,7 @@ import com.jogamp.opengl.GLEventListener;
 import com.jogamp.opengl.glu.GLU;
 import com.willwinder.ugs.nbm.visualizer.options.VisualizerOptions;
 import com.willwinder.ugs.nbm.visualizer.renderables.Grid;
-import com.willwinder.ugs.nbm.visualizer.renderables.Machine;
+import com.willwinder.ugs.nbm.visualizer.renderables.MachineBoundry;
 import com.willwinder.ugs.nbm.visualizer.renderables.MouseOver;
 import com.willwinder.ugs.nbm.visualizer.renderables.OrientationCube;
 import com.willwinder.ugs.nbm.visualizer.renderables.Tool;
@@ -137,7 +137,7 @@ public class GcodeRenderer implements GLEventListener, IRenderableRegistrationSe
         setHorizontalTranslationVector();
 
         objects = new ArrayList<>();
-        objects.add(new Machine());
+        objects.add(new MachineBoundry());
         objects.add(new Tool(Localization.getString("platform.visualizer.renderable.tool-location")));
         objects.add(new MouseOver(Localization.getString("platform.visualizer.renderable.mouse-indicator")));
         objects.add(new OrientationCube(0.5f, Localization.getString("platform.visualizer.renderable.orientation-cube")));
